add dark mode handling to network stargazers plot
